Change X.L.StateFull semantics for multi-monitor
The "last true focus" previously tracked by X.L.StateFull.FocusTracking
was the /WindowSet/ focus. With the intention of improving the
multi-monitor experience, it now tracks /WindowSpace/ focus instead.

That said, I was actually using two or three monitors when I originally
wrote this module, so it may not be mere oversight---I have a vague
feeling I used the WindowSet focus for a reason. If so, that reason is
lost to oblivion; we'll have to rediscover it.